Finish file split
