+++ exifread/__init__.py
@@ -4,7 +4,8 @@ Read Exif metadata from tiff and jpeg files.
 from typing import BinaryIO
 
 from .exif_log import get_logger
-from .classes import ExifHeader, Ifd
+from .exif_header import ExifHeader
+from .ifd import Ifd
 from .tags import DEFAULT_STOP_TAG
 from .utils import determine_type
 from .heic import HEICExifFinder
